requirements.txt
# NotesGPT Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Templates
jinja2==3.1.3

# LangChain Core
langchain
langchain-community
langchain-groq
langchain-huggingface

# Vector Store
langchain-chroma
chromadb

# Embeddings (HuggingFace)
sentence-transformers==2.7.0

# Document Processing
pypdf==4.0.1

# Environment Variables
python-dotenv==1.0.1
